[jolla-notes] Fix notes cover labels getting clipped on Qt 5. Contributes to JB#7025
On Qt 5 OpacityRampEffect clipped note text on covers, because
the text item height was smaller than the implicit content height.

Also on Qt 5 ListView.clip caused the first cover label delegate
to clip it's ascenders, avoided this with header item.

In the design page numbers should align to the last line of text,
which was chosen with ad-hoc parameters on Qt 4, now with Qt 5 we
can use lineHeight property to calculate the exact position.

Trim text so that trailing empty lines don't reserve space on the cover.