bugfix: create RemorseItem with the delegate as its parent, contributes JB#17563
Deleting a note will also cause the delegate to be deleted, which cleans
up the RemorseItem automatically if the RemorseItem is a child of the
delegate. This is better than the previous situation with the RemorseItem
as a child of the view, which would leave it with a dangling reference
to an already-destroyed delegate.