[tests] Separate fixture-creation from the testcase files
This separation allows the db setup (creating a handful of notes for
use in testing) to be run before the app starts, which is (a) more
accurate and (b) faster, because the setup can be done once and then reused.

Since this makes invocation of the tests more complicated, added a
runtest script to help. It also means that the whole directory can't
be run at once via qmltestrunner, so tests.xml has been changed to
call runtest for each test file.