[jolla-notes] Use default vNote encoding if possible. Contributes to JB#16502
This commit checks whether the note string can be displayed using
only us-ascii characters, and if so uses the default encoding and
character set.

If non-ascii characters are present, this commit now ensures that
we use CHARSET=UTF-8 with ENCODING=8BIT instead of using
ENCODING=BASE64 as we did previously.

It also checks whether any linefeed chars occur in the string, as
section 10.7.3 of IrMC states that the body field of the vNote can
contain any character from the default encoding character set
except for linefeed.  Thus, we don't do "normal" folding, but
instead replace all linefeed characters with carriage returns.

Finally, it converts the timestamps into UTC timespec, and ensures
that the stringified versions are formatted in the expected way.

In general, this commit will ensure that most notes will be human
readable after sharing via bluetooth or email.

Contributes to JB#16502