[vault] Fix pkill commandline arguments. JB#48950
pkill takes a pattern as the commandline argument.

/usr/libexec/jolla-notes/notes-vault --action export --name Notes [...]
which runs the pkill process kills both the /usr/bin/jolla-notes proces
and the parent itself as well, if run under pkill from busybox.

It seems there is some difference in behaviour against procps-ng, in
which pkill from there does not match against directory name in the
command name.

Making it explicit that the argument is a pattern, anchored at the end,
fixes the problem. I also added the beginning anchor, in case there is a
process in the future that would be named "this-is-not-jolla-notes".

I found that neither procps-ng nor busybox has a problem matching the
command with both beginning and end anchor, even when they are listed
with their pathnames in process list.